Put clustering functions into another module
